Removed activation driver api
Because of many cyclic dependencies with
activation-driver/user-provider/repositories the activation part was
completely rewritten.
Now the Registrar and many other parts rely on a Token\Repository, which
stores tokens for activation/remembering/password-resets and so on.

The user has to be activatable to can be activated. I prevent
implementations on value objects like the user object at almost any
cost, but the activation is much simpler with it.