Implemented a new default Permission Checker
The login as function did not work, because requiredPermissionCodes()
of the other user returns nothing which leads to an access denied for
the user trying to login as the other.
The new version merges all codes and the new checker checks on this
merged permissions.

CAUTION: Permission wildcard support has been removed! The exact
inheritance and priority of wildcard permissions are nearly impossible
to ensure. Look at tests if you dont believe